puesta al dia

+++ nowplaying/api/database.py
@@ -4,7 +4,7 @@ from datetime import datetime, timezone
 
 DB_FILE = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "current_track.json"))
 DB_HISTORY_FILE = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "history.json"))
-MARGEN_MINUTOS = 2
+MARGEN_MINUTOS = 1
 
 # ==========================================
 # GESTIÓN DE LA CANCIÓN ACTUAL
@@ -14,12 +14,41 @@ def save_current_track(track_data):
     """Guarda la canción o el fallo. Si es un éxito, lo añade al historial."""
     track_data["timestamp"] = datetime.now(timezone.utc).isoformat()
     
-    with open(DB_FILE, "w", encoding="utf-8") as f:
-        json.dump(track_data, f, indent=4, ensure_ascii=False)
-        
-    # 🆕 Si la canción ha sido un éxito, intentamos guardarla en el historial
+
     if track_data.get("success") is True:
+        with open(DB_FILE, "w", encoding="utf-8") as f:
+            json.dump(track_data, f, indent=4, ensure_ascii=False)
         add_to_history(track_data)
+        return
+
+    if not os.path.exists(DB_FILE):
+        with open(DB_FILE, "w", encoding="utf-8") as f:
+            json.dump(track_data, f, indent=4, ensure_ascii=False)
+        return
+
+    # Leemos la última canción registrada
+    with open(DB_FILE, "r", encoding="utf-8") as f:
+        data_actual = json.load(f)
+
+    # Si lo que hay guardado ya era un fallo, no hacemos nada
+    if data_actual.get("success") is False:
+        return
+
+    # Si hay una canción real en el JSON, medimos cuánto tiempo lleva sonando
+    last_time_str = data_actual.get("timestamp")
+    if last_time_str:
+        last_time = datetime.fromisoformat(last_time_str)
+        ahora = datetime.now(timezone.utc)
+        diferencia = (ahora - last_time).total_seconds() / 60
+
+        # Solo marcamos como Unknown si ha pasado el tiempo de gracia
+        if diferencia > MARGEN_MINUTOS:
+            with open(DB_FILE, "w", encoding="utf-8") as f:
+                json.dump(track_data, f, indent=4, ensure_ascii=False)
+
+
+
+
 
 def get_current_track():
     """Lee la canción actual y aplica el margen si no se detecta nada"""
@@ -37,20 +66,7 @@ def get_current_track():
     with open(DB_FILE, "r", encoding="utf-8") as f:
         data = json.load(f)
 
-    if data.get("success") is True:
-        return data
 
-    last_time_str = data.get("timestamp")
-    if not last_time_str:
-        return estado_vacio
-
-    last_time = datetime.fromisoformat(last_time_str)
-    ahora = datetime.now(timezone.utc)
-    diferencia = (ahora - last_time).total_seconds() / 60
-
-    if diferencia > MARGEN_MINUTOS:
-        return estado_vacio
-        
     return data
 
 # ==========================================
